Refactor memory management and improve performance in CachingAllocator
- Removed block splitting logic in CachingAllocator::find_free_block to simplify memory management and avoid sub-pointer issues.
- Updated MaskRCNN to correctly calculate the number of anchors based on scales and aspect ratios.
- Introduced AttentionTypeCastBackward for autograd-aware dtype casting in attention layers, ensuring proper gradient flow.
- Enhanced BatchNorm layers to safely increment num_batches_tracked across devices.
- Streamlined Conv layers to use OpId-based dispatch for backward operations, improving clarity and performance.
- Updated normalization layers to handle dtype conversions for backward computations, preventing gradient overflow.
- Modified pooling layers to save output tensors for cuDNN backward compatibility.
- Adjusted tensor creation functions to ensure correct attribute settings.
- Improved error handling in broadcasting operations by using runtime exceptions.
- Updated unit tests for MaskRCNN and additional operations to ensure dtype compatibility and correct behavior.
- Adjusted transform tests to align with PyTorch behavior regarding tensor squeezing.
- Added warmup pass in YOLO tests to ensure BatchNorm statistics are populated before evaluation.

+++ tests/unit/test_ops_additional_multidtype.cpp
@@ -133,7 +133,8 @@ TEST_P(OpsAdditionalMultiDTypeTest, SumAllElements) {
     } else if (dtype == DType::Float64) {
         EXPECT_DOUBLE_EQ(result_cpu.data<double>()[0], 24.0);
     } else if (dtype == DType::Int32) {
-        EXPECT_EQ(result_cpu.data<int32_t>()[0], 24);
+        // Int32 sum promotes to Int64 to prevent overflow (matches PyTorch)
+        EXPECT_EQ(result_cpu.data<int64_t>()[0], 24);
     }
 }
 
@@ -260,7 +261,8 @@ TEST_P(OpsAdditionalMultiDTypeTest, ProdOperation) {
     if (dtype == DType::Float32) {
         EXPECT_FLOAT_EQ(result_cpu.data<float>()[0], 64.0f);
     } else if (dtype == DType::Int32) {
-        EXPECT_EQ(result_cpu.data<int32_t>()[0], 64);
+        // Int32 prod promotes to Int64 to prevent overflow (matches PyTorch)
+        EXPECT_EQ(result_cpu.data<int64_t>()[0], 64);
     }
 }
